Begin conversion to celery
This does much of the dirty work to strip out the old scheduler and
replace it with celery scheduled tasks. Currently outgoing mail is
broken -- the next task will be to switch to using the Django email
backend rather than rolling our own.

This also cleans up some DRY problems, some places where we were being
un-Pythonic, and removes a bunch of unused imports when I was in
particular files.