Fix sending bug in mailing lists
We were treating this like a regular email, with all of the recipients
in the 'to' line. Treat it like a mailing list, with all the recipients
as BCC.

There is still some question whether or not we should send an
individual mail for each recipient, and have their email be the 'to'
address. In addition, I don't know if fail_silently will cause all
recipients to fail if one fails.