Clean up email testing, propagate threading headers, clean up subject
The major change here was to remove the TestMailChecker and just create
the IncomingMail objects directly in the unit tests. Moving the email
parsing and IncomingMail construction code into the MailingList object
allows us to actually test mail parsing.

What this allows is to test that certain threading headers (Message-ID,
References, and In-Reply-To) are propagated into the destination
message.

In addition, replies to the mailing list don't prefix the subject
*again*, which allows better threading of email messages in mail
clients that thread based on subject, not on the threading headers.