Moving settings into a Python package
Details:
https://speakerdeck.com/jacobian/the-best-and-worst-of-django?slide=81

This allows us to define multiple settings configurations as modules that can
be selected depending on the deployment environment. This provides us
with ability to load LDAP dependencies & configuration only on Nadine
deployments that require it.

You can explicity specify which settings module you want to load with
the DJANGO_SETTINGS_MODULE environment variable.