Add Stripe integration for collecting billing info
Refs #386.

Implemented with a dedicated view class; a subclass of Django's BaseFormView.
The view class is deliberately limited to only handle POST requests as
GET/render is handled elsewhere.

Added a new model (with migration), StripeBillingProfile to keep reference to a
user's Stripe Customer reference if one exists.

New code & changes are spread across nadine and staff apps as a matter
of convenience. Need to review and determine whether this feature needs
to be better isolated from rest of system (eg: an optional integration similar
to LDAP).